+++ src/PhpBrew/Command/EnvCommand.php
@@ -12,19 +12,23 @@ class EnvCommand extends \CLIFramework\Command
     public function execute($version = null)
     {
         // get current version
-        if( ! $version )
+        if ( ! $version ) {
             $version = getenv('PHPBREW_PHP');
+        }
 
         // $currentVersion;
         $root = Config::getPhpbrewRoot();
         $home = Config::getPhpbrewHome();
         $buildDir = Config::getBuildDir();
+        $lookup = getenv('PHPBREW_LOOKUP_PREFIX');
 
         // $versionBuildPrefix = Config::getVersionBuildPrefix($version);
         // $versionBinPath     = Config::getVersionBinPath($version);
 
-        echo 'export PHPBREW_ROOT=' . $root . "\n";
-        echo 'export PHPBREW_HOME=' . $home . "\n";
+        echo "export PHPBREW_ROOT=$root\n";
+        echo "export PHPBREW_HOME=$home\n";
+        echo "export PHPBREW_LOOKUP_PREFIX=$lookup\n";
+        
 
         if ($version !== false) {
             // checking php version exists